Remove ops compat indirection from wrappers
