+++ src/rook/utils/average_utils.py
@@ -1,16 +1,16 @@
 def run_average_by_time(args):
-    from .ops.compat import run_average_time
+    from .ops.average import average_time
 
-    return run_average_time(args)
+    return average_time(**args).file_uris
 
 
 def run_average_by_dim(args):
-    from .ops.compat import run_average_over_dims
+    from .ops.average import average_over_dims
 
-    return run_average_over_dims(args)
+    return average_over_dims(**args).file_uris
 
 
 def run_average_by_shape(args):
-    from .ops.compat import run_average_shape
+    from .ops.average import average_shape
 
-    return run_average_shape(args)
+    return average_shape(**args).file_uris
