+++ src/rook/utils/regrid_utils.py
@@ -1,5 +1,5 @@
 from rook.utils.input_utils import parse_custom_grid
-from rook.utils.ops.compat import run_regrid
+from rook.utils.ops.regrid import regrid as ops_regrid
 
 
 def run_regrid(args):
@@ -10,4 +10,4 @@ def run_regrid(args):
         # parse the string into a tuple/list
         args["grid"] = parse_custom_grid(args.pop("custom_grid"))
 
-    return run_regrid(args)
+    return ops_regrid(**args).file_uris
